[reflection] Apply dotnet/corefx 33101 "Consolidate AssemblyTests, AssemblyNameTests, MethodInfoTests fixtures within System.Reflection location".
