+++ src/System.Reflection/tests/AssemblyTests.netcoreapp.cs
