+++ src/System.Reflection/tests/ForwardedTypesAssembly/ForwardedTypesAssembly.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <ProjectGuid>{FA12CC14-C4A5-495D-B7AA-651E4BCC1027}</ProjectGuid>
+    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    <Configurations>netstandard-Debug;netstandard-Release</Configurations>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="ForwardedTypesAssembly.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="$(CommonPath)\System\Diagnostics\CodeAnalysis\ExcludeFromCodeCoverageAssemblyAttribute.cs">
+      <Link>Common\System\Diagnostics\CodeAnalysis\ExcludeFromCodeCoverageAssemblyAttribute.cs</Link>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\UnloadableAssembly\UnloadableAssembly.csproj">
+      <Project>{7FFD1B55-D69A-4469-B775-6CBEB1CE50B0}</Project>
+      <Name>UnloadableAssembly</Name>
+    </ProjectReference>
+  </ItemGroup>
+</Project>
