+++ src/System.Reflection/tests/Resources/EmbeddedTextFile.txt
@@ -1,0 +1,1 @@
+Embedded Text File.
