+++ src/System.Reflection/tests/UnloadableAssembly/UnloadableAssembly.cs
@@ -1,0 +1,11 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Runtime.CompilerServices;
+
+[assembly:ReferenceAssembly]
+
+public static class TypeInUnloadableAssembly {}
+
