+++ src/System.Reflection/tests/UnloadableAssembly/UnloadableAssembly.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <ProjectGuid>{7FFD1B55-D69A-4469-B775-6CBEB1CE50B0}</ProjectGuid>
+    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    <Configurations>netstandard-Debug;netstandard-Release</Configurations>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="UnloadableAssembly.cs" />
+  </ItemGroup>
+</Project>
