+++ src/System.Runtime/tests/System/DelegateTests.cs
@@ -686,7 +686,7 @@ namespace System.Tests
         public static void CreateDelegate2_Target_GenericTypeParameter()
         {
 
-            Type theT = typeof(G<>).GetTypeInfo().GenericTypeParameters[0];
+            Type theT = typeof(DummyGenericClassForDelegateTests<>).GetTypeInfo().GenericTypeParameters[0];
             Type delegateType = typeof(Func<object, object, bool>);
             AssertExtensions.Throws<ArgumentException>("target", () => Delegate.CreateDelegate(delegateType, theT, "ReferenceEquals"));
         }
@@ -1199,3 +1199,5 @@ namespace System.Tests
         #endregion Test Setup
     }
 }
+
+internal class DummyGenericClassForDelegateTests<T> { }
